requirements.txt
Flask>=2.0
requests>=2.0
looker-sdk>=1.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
